scripts/requirements.txt
pandas
openpyxl
requests
PyMuPDF
Pillow
imagehash